Use floats in Filter

+++ include/Filter.hpp
@@ -12,7 +12,7 @@ public:
 	static int constexpr N_st = 6; // Dimension of state
 	static int constexpr N_ob = 6; // Dimension of observation
 	//ctor
-	Filter(const MatrixXd<N_st> &P0, const MatrixXd<N_st> &Q0, const MatrixXd<N_ob> &R0);
+	Filter(const MatrixXf<N_st> &P0, const MatrixXf<N_st> &Q0, const MatrixXf<N_ob> &R0);
 
 	bool ini = false;
 	bool loss = false;
@@ -21,14 +21,14 @@ public:
 	const static int M = 6;
 	unsigned int F_count = 0;
 
-	std::vector<std::pair<double, Eigen::Vector3d>> position_data;
+	std::vector<std::pair<float, Eigen::Vector3f>> position_data;
 
-	MatrixXd<N_st> F, P, Q, P_pre, P_temp; //state transformtion matrix and covariance matrix and process noise
-	Matrix<double, N_ob, N_st> H;		   //observation matrix
-	Matrix<double, N_st, N_ob> K;		   //gain
-	MatrixXd<M> R;						   // covariance of observation
-	VectorXd<M> Z;						   // observation
-	VectorXd<N_st> X, X_pre, X_temp;
+	MatrixXf<N_st> F, P, Q, P_pre, P_temp; //state transformtion matrix and covariance matrix and process noise
+	Matrix<float, N_ob, N_st> H;		   //observation matrix
+	Matrix<float, N_st, N_ob> K;		   //gain
+	MatrixXf<M> R;						   // covariance of observation
+	VectorXf<M> Z;						   // observation
+	VectorXf<N_st> X, X_pre, X_temp;
 
 	void receivedata(const geometry_msgs::PoseStamped::ConstPtr &msg);
 
@@ -40,7 +40,7 @@ public:
 	
 	void compute_obsv_Jacobian();
 
-	void update(Eigen::Matrix<double, N_ob, 1> &observe);
+	void update(Eigen::Matrix<float, N_ob, 1> &observe);
 };
 
 //Publish message if predict and update has achieved a new message the flag will be 1
